balancer/requirements.txt
aiohttp==3.8.3; python_version >= "3.6"
aiosignal==1.3.1; python_version >= "3.7"
alembic==1.8.1; python_version >= "3.7"
anyio==3.6.2; python_full_version >= "3.6.2" and python_version >= "3.7"
async-timeout==4.0.2; python_version >= "3.6"
asyncpg==0.27.0; python_version >= "3.7" and python_full_version >= "3.7.0"
attrs==22.2.0; python_version >= "3.6"
bcrypt==4.0.1; python_version >= "3.6"
bidict==0.22.1; python_version >= "3.7"
charset-normalizer==2.1.1; python_full_version >= "3.6.0" and python_version >= "3.6"
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.7.0" and platform_system == "Windows"
databases==0.7.0; python_version >= "3.7"
dnspython==2.3.0; python_version >= "3.7" and python_version < "4.0"
ecdsa==0.18.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
eventlet==0.33.3
fastapi-socketio==0.0.9
fastapi==0.89.1; python_version >= "3.7"
frozenlist==1.3.3; python_version >= "3.7"
greenlet==2.0.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and python_full_version >= "3.5.0"
h11==0.14.0; python_version >= "3.7"
httptools==0.5.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.4; python_full_version >= "3.6.2" and python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
mako==1.2.4; python_version >= "3.7"
markupsafe==2.1.2; python_version >= "3.7"
multidict==6.0.4; python_version >= "3.7"
passlib==1.7.4
pillow==9.4.0; python_version >= "3.7"
psycopg2==2.9.5; python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pydantic==1.10.4; python_version >= "3.7"
python-dotenv==0.21.1; python_version >= "3.7"
python-engineio==4.3.4; python_version >= "3.6"
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.7.2; python_version >= "3.6"
pyyaml==6.0; python_version >= "3.7"
rsa==4.9; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
sqlalchemy==1.4.46; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
starlette==0.22.0; python_version >= "3.7"
typing-extensions==4.4.0; python_version >= "3.7"
uvicorn==0.20.0; python_version >= "3.7"
uvloop==0.17.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchfiles==0.18.1; python_version >= "3.7"
websockets==10.4; python_version >= "3.7"
yarl==1.8.2; python_version >= "3.7"
